+++ packages/web-components/src/tree-item/tree-item.base.ts
@@ -114,6 +114,9 @@ export class BaseTreeItem extends FASTElement {
       return;
     }
 
+    //If a tree item is nested and initially set to selected expand the tree items so the selected item is visible
+    this.expanded = Array.from(this.querySelectorAll('*')).some(el => isTreeItem(el) && el.selected);
+
     this.childTreeItems.forEach(item => {
       this.setIndent(item);
     });
@@ -128,33 +131,6 @@ export class BaseTreeItem extends FASTElement {
   }
 
   /**
-   * Handle focus events
-   *
-   * @public
-   */
-  public focusHandler(e: FocusEvent): void {
-    if (
-      e.target === this ||
-      // In case where the tree-item contains a focusable element, we should not set the tabindex to 0 when the focus is on its child focusable element,
-      // so users can shift+tab to navigate to the tree-item from its child focusable element.
-      this.contains(e.target as Node)
-    ) {
-      this.setAttribute('tabindex', '0');
-    }
-  }
-
-  /**
-   * Handle blur events
-   *
-   * @public
-   */
-  public blurHandler(e: FocusEvent): void {
-    if (e.target === this) {
-      this.setAttribute('tabindex', '-1');
-    }
-  }
-
-  /**
    * Toggle the expansion state of the tree item
    *
    * @public
@@ -163,15 +139,6 @@ export class BaseTreeItem extends FASTElement {
     if (this.childTreeItems?.length) {
       this.expanded = !this.expanded;
     }
-  }
-
-  /**
-   * Toggle the single selection state of the tree item
-   *
-   * @public
-   */
-  public toggleSelection() {
-    this.selected = !this.selected;
   }
 
   /**